refactor: Extract Statistics page components
