+++ app/pages/statistics.tsx
@@ -1,31 +1,17 @@
-/* eslint-disable visualize-admin/no-large-sx */
-import { Box, Card, Link, Tooltip, Typography } from "@mui/material";
-import { max, rollups, sum } from "d3-array";
-import { formatLocale } from "d3-format";
-import { motion } from "framer-motion";
-import uniq from "lodash/uniq";
+import { Box } from "@mui/material";
+import { rollups, sum } from "d3-array";
 import { GetServerSideProps } from "next";
-import { ComponentProps, ReactNode, useMemo } from "react";
 
 import { AppLayout } from "@/components/layout";
 import { BANNER_MARGIN_TOP } from "@/components/presence";
 import prisma from "@/db/client";
 import { deserializeProps, Serialized, serializeProps } from "@/db/serialize";
-import { useFlag } from "@/flags";
 import { Icon } from "@/icons";
-import { Locale } from "@/locales/locales";
 import { useLocale } from "@/src";
-
-type StatProps = {
-  countByDay: {
-    day: Date;
-    count: number;
-  }[];
-  trendAverages: {
-    lastMonthDailyAverage: number;
-    previousThreeMonthsDailyAverage: number;
-  };
-};
+import { BaseStatsCard } from "@/statistics/base-stats-card";
+import { ChartLink } from "@/statistics/chart-link";
+import { formatInteger } from "@/statistics/formatters";
+import { StatProps, StatsCard } from "@/statistics/stats-card";
 
 type PageProps = {
   charts: StatProps & {
@@ -333,388 +319,4 @@ const Statistics = (props: Serialized<PageProps>) => {
   );
 };
 
-const ChartLink = ({
-  locale,
-  chartKey,
-}: {
-  locale: Locale;
-  chartKey: string;
-}) => {
-  return (
-    <Link
-      href={`/${locale}/v/${chartKey}`}
-      target="_blank"
-      color="primary"
-      sx={{
-        display: "flex",
-        gap: 2,
-        justifyContent: "space-between",
-      }}
-    >
-      {chartKey}
-      <Icon name="linkExternal" size={16} />
-    </Link>
-  );
-};
-
 export default Statistics;
-
-const formatYearMonth = (date: Date, { locale }: { locale: Locale }) => {
-  const year = date.getFullYear();
-  const month = date.toLocaleDateString(locale, { month: "short" });
-  return `${year} ${month}`;
-};
-
-const groupByYearMonth = (
-  countByDay: { day: Date; count: number }[],
-  { locale }: { locale: Locale }
-) => {
-  const countByDate = rollups(
-    countByDay,
-    (v) => ({
-      count: sum(v, (d) => d.count),
-      date: v[v.length - 1].day,
-      label: formatYearMonth(v[0].day, { locale }),
-    }),
-    (d) => formatYearMonth(d.day, { locale })
-  ).reverse();
-  const allYearMonthStrings = uniq(
-    countByDate.map(([yearMonthStr]) => yearMonthStr)
-  );
-  const start = new Date(countByDate[0][1].date);
-  const end = new Date(countByDate[countByDate.length - 1][1].date);
-  if (start.getTime() !== end.getTime()) {
-    for (let date = start; date <= end; date.setMonth(date.getMonth() + 1)) {
-      const formattedDate = formatYearMonth(date, { locale });
-      if (!allYearMonthStrings.includes(formattedDate)) {
-        countByDate.push([
-          formattedDate,
-          {
-            count: 0,
-            date: new Date(date),
-            label: formattedDate,
-          },
-        ]);
-      }
-    }
-  }
-  countByDate.sort(([_a, a], [_b, b]) => b.date.getTime() - a.date.getTime());
-  return countByDate;
-};
-
-const BaseStatsCard = ({
-  title,
-  subtitle,
-  data,
-  columnName = "Date",
-  trend,
-  showPercentage = false,
-}: {
-  title: string;
-  subtitle: string;
-  data: [string, { count: number; label: ReactNode }][];
-  columnName?: string;
-  trend?: {
-    direction: "up" | "down";
-    lastMonthDailyAverage: number;
-    previousThreeMonthsDailyAverage: number;
-  };
-  showPercentage?: boolean;
-}) => {
-  const totalCount = sum(data, ([, { count }]) => count) ?? 1;
-  const maxCount = max(data, ([, { count }]) => count) ?? 1;
-
-  return (
-    <Card
-      sx={{
-        width: "100%",
-        pt: 4,
-        boxShadow: 2,
-        borderRadius: 4,
-        borderBottomLeftRadius: 0,
-        borderBottomRightRadius: 0,
-        backgroundColor: "grey.200",
-      }}
-    >
-      <Box sx={{ px: 5 }}>
-        <Box sx={{ display: "grid", gridTemplateColumns: "1fr 24px", gap: 2 }}>
-          <Typography variant="h2" sx={{ fontWeight: "normal" }}>
-            {title}
-          </Typography>
-          {trend ? (
-            <Tooltip
-              title={
-                <>
-                  <Typography variant="h3" sx={{ fontWeight: "bold" }}>
-                    {trend.direction === "up"
-                      ? "Upward trend"
-                      : "Downward trend"}
-                  </Typography>
-                  <Typography variant="caption">
-                    Last 30 days daily average:{" "}
-                    <b>
-                      {trend.lastMonthDailyAverage >= 10
-                        ? formatInteger(trend.lastMonthDailyAverage)
-                        : trend.lastMonthDailyAverage.toFixed(2)}
-                    </b>
-                  </Typography>
-                  <br />
-                  <Typography variant="caption">
-                    Last 90 days daily average:{" "}
-                    <b>
-                      {trend.previousThreeMonthsDailyAverage >= 10
-                        ? formatInteger(trend.previousThreeMonthsDailyAverage)
-                        : trend.previousThreeMonthsDailyAverage.toFixed(2)}
-                    </b>
-                  </Typography>
-                </>
-              }
-            >
-              <Typography variant="h4" component="span" sx={{ mt: "0.5em" }}>
-                {trend.direction === "up" ? "📈" : "📉"}
-              </Typography>
-            </Tooltip>
-          ) : null}
-        </Box>
-        <Typography variant="h3" sx={{ fontWeight: "normal" }}>
-          {subtitle}
-        </Typography>
-      </Box>
-      <Box
-        sx={{
-          display: "grid",
-          gridTemplateColumns: showPercentage
-            ? "auto auto auto 1fr"
-            : "auto auto 1fr",
-          gridTemplateRows: "auto",
-          gap: "2px",
-          width: "100%",
-          height: "fit-content",
-          // Roughly 12 rows
-          maxHeight: "calc(12 * 2.3rem)",
-          overflowY: "auto",
-          mt: 3,
-          p: 1,
-          pt: 0,
-          backgroundColor: "grey.200",
-        }}
-      >
-        <Box
-          sx={{
-            position: "sticky",
-            top: 0,
-            width: "100%",
-            py: "0.3rem",
-            px: "0.6rem",
-            borderBottom: (t) => `2px solid ${t.palette.divider}`,
-            backgroundColor: "background.paper",
-          }}
-        >
-          <Typography
-            variant="caption"
-            sx={{ position: "sticky", top: 0, fontWeight: "bold" }}
-          >
-            {columnName}
-          </Typography>
-        </Box>
-        <Box
-          sx={{
-            position: "sticky",
-            top: 0,
-            width: "100%",
-            py: "0.3rem",
-            px: "0.6rem",
-            borderBottom: (t) => `2px solid ${t.palette.divider}`,
-            backgroundColor: "background.paper",
-            textAlign: "end",
-          }}
-        >
-          <Typography variant="caption" sx={{ fontWeight: "bold" }}>
-            Count
-          </Typography>
-        </Box>
-        {showPercentage ? (
-          <Box
-            sx={{
-              position: "sticky",
-              top: 0,
-              width: "100%",
-              py: "0.3rem",
-              px: "0.6rem",
-              borderBottom: (t) => `2px solid ${t.palette.divider}`,
-              backgroundColor: "background.paper",
-              textAlign: "end",
-            }}
-          >
-            <Typography variant="caption" sx={{ fontWeight: "bold" }}>
-              Share
-            </Typography>
-          </Box>
-        ) : null}
-        <Box
-          sx={{
-            zIndex: 1,
-            position: "sticky",
-            top: 0,
-            width: "100%",
-            height: "100%",
-            p: "0.3rem",
-            borderBottom: (t) => `2px solid ${t.palette.divider}`,
-            backgroundColor: "background.paper",
-          }}
-        />
-        {data.map(([key, datum]) => (
-          <Bar
-            key={key}
-            {...datum}
-            maxCount={maxCount}
-            totalCount={totalCount}
-            showPercentage={showPercentage}
-          />
-        ))}
-      </Box>
-    </Card>
-  );
-};
-
-const Bar = ({
-  label,
-  count,
-  maxCount,
-  totalCount,
-  showPercentage,
-}: ComponentProps<typeof BaseStatsCard>["data"][number][1] & {
-  label: ReactNode;
-  maxCount: number;
-  totalCount: number;
-  showPercentage?: boolean;
-}) => {
-  const easterEgg = useFlag("easter-eggs");
-
-  return (
-    <>
-      <Box
-        sx={{
-          width: "100%",
-          py: "0.3rem",
-          px: "0.6rem",
-          backgroundColor: "background.paper",
-        }}
-      >
-        <Typography variant="caption" sx={{ cursor: "default" }}>
-          {label}
-        </Typography>
-      </Box>
-      <Box
-        sx={{
-          width: "100%",
-          py: "0.3rem",
-          px: "0.6rem",
-          backgroundColor: "background.paper",
-          textAlign: "end",
-        }}
-      >
-        <Typography variant="caption">{formatInteger(count)}</Typography>
-      </Box>
-      {showPercentage ? (
-        <Box
-          sx={{
-            width: "100%",
-            py: "0.3rem",
-            px: "0.6rem",
-            backgroundColor: "background.paper",
-            textAlign: "end",
-          }}
-        >
-          <Typography variant="caption">
-            {formatInteger((count / totalCount) * 100)}%
-          </Typography>
-        </Box>
-      ) : null}
-      <Box
-        sx={{
-          width: "100%",
-          height: "100%",
-          p: "0.3rem",
-          backgroundColor: "background.paper",
-        }}
-      >
-        {count > 0 ? (
-          !easterEgg ? (
-            <Box
-              sx={{
-                width: `${(count / maxCount) * 100}%`,
-                height: "100%",
-                backgroundColor: "primary.main",
-              }}
-            />
-          ) : (
-            <>
-              {Array(Math.ceil((count / maxCount) * 40))
-                .fill(0)
-                .map((_, i) => {
-                  return (
-                    <motion.div
-                      key={i}
-                      style={{ display: "inline-block" }}
-                      initial={{ scale: 0, opacity: 0 }}
-                      animate={{
-                        opacity: 1,
-                        scale: 1,
-                        transition: { delay: i * 0.03 },
-                      }}
-                    >
-                      🥓
-                    </motion.div>
-                  );
-                })}
-            </>
-          )
-        ) : null}
-      </Box>
-    </>
-  );
-};
-
-const StatsCard = (
-  props: StatProps & {
-    title: (total: number) => string;
-    subtitle: (total: number, avgMonthlyCount: number) => string;
-  }
-) => {
-  const { title, subtitle, countByDay, trendAverages } = props;
-  const locale = useLocale();
-  const { countByYearMonth, total } = useMemo(() => {
-    return {
-      countByYearMonth: groupByYearMonth(countByDay, { locale }),
-      total: sum(countByDay, (d) => d.count) ?? 0,
-    };
-  }, [countByDay, locale]);
-  const avgMonthlyCount = Math.round(total / countByYearMonth.length);
-  const { lastMonthDailyAverage, previousThreeMonthsDailyAverage } =
-    trendAverages;
-  return (
-    <BaseStatsCard
-      title={title(total)}
-      subtitle={subtitle(total, avgMonthlyCount)}
-      data={countByYearMonth}
-      trend={{
-        direction:
-          lastMonthDailyAverage > previousThreeMonthsDailyAverage
-            ? "up"
-            : "down",
-        lastMonthDailyAverage,
-        previousThreeMonthsDailyAverage,
-      }}
-    />
-  );
-};
-
-const formatInteger = formatLocale({
-  decimal: ".",
-  thousands: "\u00a0",
-  grouping: [3],
-  currency: ["", "\u00a0 CHF"],
-  minus: "\u2212",
-  percent: "%",
-}).format(",d");
